src/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
kafka-python==2.0.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
prometheus-client==0.19.0
requests==2.31.0
